+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Misc:
 - **Misspell** Support `ignore` option as array [#1255](https://github.com/sider/runners/pull/1255)
 - Allow `location(start_line,start_column)` format [#1257](https://github.com/sider/runners/pull/1257)
 - **ESLint** Add `target` option instead of `dir` option [#1264](https://github.com/sider/runners/pull/1264)
+- **HAML-Lint** Add `target` option [#1265](https://github.com/sider/runners/pull/1265)
 
 ## 0.29.3